Split modal upload form into a few different files, following conventions of other forms across the app.
