+++ garden/src/components/form/modal/create/ModalUploadPage.tsx
@@ -2,7 +2,7 @@ import { ModalUploadForm } from "./ModalUploadForm";
 
 export const ModalUploadPage = () => {
   return (
-    <div className="mx-auto max-w-5xl p-20 font-display">
+    <div className="mx-auto max-w-4xl py-12 font-display">
       <ModalUploadForm />
     </div>
   );
